# website/src/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "@mintopia/harmonic",
    "version": "1.0.0",
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A key created via POST /api/keys, sent as `Authorization: Bearer <token>`."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "harmonic_session",
        "description": "The session cookie set by POST /api/auth/login."
      }
    },
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string",
                "example": "not_found"
              },
              "message": {
                "type": "string",
                "example": "no task with id 4821"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "error"
        ],
        "additionalProperties": false
      },
...
      "Workspace": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 1
          },
          "name": {
            "type": "string",
            "example": "Harmonic"
          },
          "workingDir": {
            "type": "string",
            "example": "/home/dev/harmonic"
          },
          "trackerEnabled": {
            "type": "boolean",
            "example": false
          },
          "trackerPollIntervalSeconds": {
            "type": "number",
            "example": 60
          },
          "resolvedTracker": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "ok": {
                    "type": "boolean",
                    "example": true
                  },
                  "label": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "example": "GitHub"
                  },
                  "code": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "example": null
                  },
                  "reason": {
                    "anyOf": [
                      {
                        "type": "string"
                      },
                      {
                        "type": "null"
                      }
                    ],
                    "example": null
                  }
                },
                "required": [
                  "ok",
                  "label",
                  "code",
                  "reason"
                ],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ],
            "description": "The tracker this Workspace resolved (issue #83), or null when tracking is off."
          },
          "harness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": null
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": null
          },
          "chatHarness": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": null
          },
          "chatModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": null
          },
          "isolationMode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": null
          },
          "priority": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "example": null
          },
          "maxConcurrentRuns": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "example": null
          },
          "autoRunnerEnabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "example": null
          },
          "createdAt": {
            "type": "number",
            "example": 1784030400000
          },
          "updatedAt": {
            "type": "number",
            "example": 1784032260000
          }
        },
        "required": [
          "id",
          "name",
          "workingDir",
          "trackerEnabled",
          "trackerPollIntervalSeconds",
          "resolvedTracker",
          "harness",
          "model",
          "chatHarness",
          "chatModel",
          "isolationMode",
          "priority",
          "maxConcurrentRuns",
          "autoRunnerEnabled",
          "createdAt",
          "updatedAt"
        ],
        "additionalProperties": false
      },
...
    }
  },
  "paths": {
...
    "/api/workspaces": {
      "get": {
        "tags": [
          "Workspaces"
        ],
        "description": "List Workspaces. Operator only; not reachable with a run-scoped Run Key.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "Every Workspace, oldest first.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "workspaces": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Workspace"
                      }
                    }
                  },
                  "required": [
                    "workspaces"
                  ],
                  "additionalProperties": false,
                  "description": "Every Workspace, oldest first."
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Workspaces"
        ],
        "description": "Create a Workspace: a named Working Directory, unique by absolute path. Operator only; not reachable with a run-scoped Run Key.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Harmonic"
                  },
                  "workingDir": {
                    "type": "string",
                    "minLength": 1,
                    "example": "/home/dev/harmonic"
                  },
                  "trackerEnabled": {
                    "example": false,
                    "type": "boolean"
                  },
                  "trackerPollIntervalSeconds": {
                    "example": 60,
                    "type": "integer",
                    "minimum": 5,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "name",
                  "workingDir"
                ]
              }
            }
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "201": {
            "description": "The created Workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "description": "The payload failed validation, or the working directory does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Another Workspace already uses that absolute path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}": {
      "get": {
        "tags": [
          "Workspaces"
        ],
        "description": "Get one Workspace. Operator only; not reachable with a run-scoped Run Key.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991,
              "example": 4821
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The Workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "404": {
            "description": "No Workspace has that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Workspaces"
        ],
        "description": "Rename a Workspace or repoint its Working Directory. Operator only; not reachable with a run-scoped Run Key.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "example": "Harmonic"
                  },
                  "workingDir": {
                    "type": "string",
                    "minLength": 1,
                    "example": "/home/dev/harmonic"
                  },
                  "trackerEnabled": {
                    "example": false,
                    "type": "boolean"
                  },
                  "trackerPollIntervalSeconds": {
                    "example": 60,
                    "type": "integer",
                    "minimum": 5,
                    "maximum": 9007199254740991
                  },
                  "harness": {
                    "example": "codex",
                    "anyOf": [
                      {
                        "type": "string",
                        "minLength": 1
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "model": {
                    "example": "gpt-5",
                    "anyOf": [
                      {
                        "type": "string",
                        "minLength": 1
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "chatHarness": {
                    "example": "codex",
                    "anyOf": [
                      {
                        "type": "string",
                        "minLength": 1
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "chatModel": {
                    "example": "gpt-5.6-sol",
                    "anyOf": [
                      {
                        "type": "string",
                        "minLength": 1
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "isolationMode": {
                    "example": "worktree",
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "direct",
                          "worktree"
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "priority": {
                    "example": "high",
                    "anyOf": [
                      {
                        "type": "string",
                        "enum": [
                          "high",
                          "normal",
                          "low"
                        ]
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "maxConcurrentRuns": {
                    "example": 2,
                    "anyOf": [
                      {
                        "type": "integer",
                        "minimum": 1,
                        "maximum": 9007199254740991
                      },
                      {
                        "type": "null"
                      }
                    ]
                  },
                  "autoRunnerEnabled": {
                    "example": true,
                    "anyOf": [
                      {
                        "type": "boolean"
                      },
                      {
                        "type": "null"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991,
              "example": 4821
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The updated Workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Workspace"
                }
              }
            }
          },
          "400": {
            "description": "The payload failed validation, or the working directory does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "No Workspace has that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Another Workspace already uses that absolute path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Workspaces"
        ],
        "description": "Delete a Workspace and everything on its board, stopping its tracker poll loop. Refuses a Workspace with a running Task; deleting the last Workspace is allowed. Operator only; not reachable with a run-scoped Run Key.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991,
              "example": 4821
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "204": {
            "description": "The Workspace and its board were deleted."
          },
          "404": {
            "description": "No Workspace has that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "It has a running Task.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{id}/tracker/refresh": {
      "post": {
        "tags": [
          "Workspaces"
        ],
        "description": "Force an immediate tracker poll for a Workspace — rescan its Working Directory and mirror any ticket changes onto the board now, instead of waiting for the next interval. Operator only; not reachable with a run-scoped Run Key.",
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": -9007199254740991,
              "maximum": 9007199254740991,
              "example": 4821
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {
            "sessionCookie": []
          }
        ],
        "responses": {
          "200": {
            "description": "The tracker was re-polled.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    }
                  },
                  "required": [
                    "ok"
                  ],
                  "additionalProperties": false,
                  "description": "The tracker was re-polled."
                }
              }
            }
          },
          "404": {
            "description": "No Workspace has that id.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Tracking is not enabled for this Workspace.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "The tracker scan failed (e.g. an unreadable ticket directory).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}